chore: Moving static layout styled components to a common file for re-using on EE (#39379)
## Description

- Moving static layout styled components to a common file for re-using
on EE
- Updating AppIDE folder name: `layout` to `layouts`
- Showing the Add button for libraries via props to re-use the component
on EE

Fixes [#39037](https://github.com/appsmithorg/appsmith/issues/39037)

## Automation

/ok-to-test tags="@tag.All"

### :mag: Cypress test results
