+++ app/client/src/ce/pages/common/AppHeader.tsx
@@ -19,7 +19,7 @@ import {
 } from "constants/routes";
 import Navigation from "pages/AppViewer/Navigation";
 import type { RouteComponentProps } from "react-router";
-import { Header as AppIDEHeader } from "pages/AppIDE/layout/components/Header";
+import { Header as AppIDEHeader } from "pages/AppIDE/layouts/components/Header";
 
 export type Props = RouteComponentProps;
