+++ app/client/src/pages/AppIDE/layouts/AnimatedLayout.tsx
