+++ app/client/src/pages/AppIDE/layouts/components/EditorTabs/ScreenModeToggle.tsx
