+++ app/client/src/pages/AppIDE/layouts/components/EditorTabs/constants.ts
