+++ app/client/src/pages/AppIDE/layouts/components/SegmentSwitcher/SegmentSwitcher.tsx
