+++ app/client/src/pages/AppIDE/layouts/components/SegmentSwitcher/useSegmentNavigation.ts
