+++ app/client/src/pages/AppIDE/layouts/routers/JSEditor/index.ts
