+++ app/client/src/pages/AppIDE/layouts/routers/QueryEditor/Editor.tsx
