+++ app/client/src/pages/AppIDE/layouts/routers/QueryEditor/index.ts
